requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.25.2
playwright==1.40.0
playwright-stealth==2.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
python-dotenv==1.0.0

